removed TitleProvider and use getPageTitle() from support library
The support library's latest revision has a method called
getPageTitle(). Now we got this there is no need for the TitleProvider
anymore. After the change the PagerAdapter doesn't need to be changed to
implement ViewPagerIndicator into the project.
The default implementation of getPageTitle() returns null. My
implementation shows an empty string if null was returned. Maybe there
is a better thing to show like the position.